requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose
passlib[bcrypt]
pydantic
